requirements.txt
streamlit==1.24.0
pandas==1.3.5
python-dotenv==0.19.0
Pillow==9.5.0 